Add cricket records section with reusable leaderboard MDX component
Extract leaderboard UI into a self-contained LeaderboardContent component
that manages state via URL search params (shareable on any page). Register
as <Leaderboard /> MDX component and create cricket/records content pages.
Old /leaderboard/:year URLs redirect to ?season= format for backwards compat.
