+++ apps/web/content/pages/cricket/records/_index.mdx
@@ -1,0 +1,7 @@
+---
+title: Records
+description: "Percy Main Cricket Club records — leaderboards, statistics, and honours."
+menuOrder: 5
+---
+
+Browse our club cricket records, statistics, and honours from over 160 years of Percy Main Cricket Club history.
