+++ apps/web/content/pages/cricket/records/leaderboards.mdx
@@ -1,0 +1,7 @@
+---
+title: Leaderboards
+description: "Percy Main Cricket Club batting and bowling leaderboards — season and all-time statistics."
+menuOrder: 1
+---
+
+<Leaderboard />
